# docs/openapi.yaml
# Здесь будет OpenAPI описание всех доступных эндпоинтов и моделей данных.
openapi: 3.1.0
info:
  title: Bank API
  description: REST API для управления банковскими картами
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
paths:
  /api/admin/users/{userId}:
    get:
      tags:
        - admin-rest-controller
      summary: Найти пользователя
      description: Найти пользователя в БД по ID
      operationId: findUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    put:
      tags:
        - admin-rest-controller
      summary: Обновить пользователя
      description: Обновить уже созданного пользователя в БД
      operationId: update
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    delete:
      tags:
        - admin-rest-controller
      summary: Удалить пользователя из БД по ID
      operationId: deleteById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
  /api/admin/cards/{cardId}:
    get:
      tags:
        - admin-rest-controller
      summary: Найти карту в БД по ID
      operationId: findCardById
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
    put:
      tags:
        - admin-rest-controller
      summary: Обновить карту в БД
      operationId: updateCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
    delete:
      tags:
        - admin-rest-controller
      summary: Удалить карту из БД
      operationId: deleteCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
  /api/admin/cards/{cardId}/block:
    put:
      tags:
        - admin-rest-controller
      summary: Блокировать карту в БД
      operationId: blockCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
  /api/admin/cards/{cardId}/activate:
    put:
      tags:
        - admin-rest-controller
      summary: Активировать карту в БД
      operationId: activateCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
  /authorization/login:
    post:
      tags:
        - auth-controller
      summary: Авторизация
      description: Для авторизации зарегистрированных пользователей
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/cards/transfer:
    post:
      tags:
        - card-rest-controller
      summary: Перевод
      description: Перевод средств между двумя картами пользователя
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionResultDTO"
  /api/admin/users:
    get:
      tags:
        - admin-rest-controller
      summary: Найти всех пользователей
      operationId: findAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
    post:
      tags:
        - admin-rest-controller
      summary: Создать нового пользователя
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
  /api/admin/users/{userId}/cards:
    post:
      tags:
        - admin-rest-controller
      summary: Добавить пользователю карту
      operationId: addUserCard
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardAddDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
  /api/admin/cards:
    get:
      tags:
        - admin-rest-controller
      summary: Получить все карты из БД
      operationId: getAllCards
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardDTO"
    post:
      tags:
        - admin-rest-controller
      summary: Создать новую карту
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
  /api/cards/{cardId}/requestBlock:
    patch:
      tags:
        - card-rest-controller
      summary: Запросить блок
      description: Метод для запроса блокировки карты по ID
      operationId: requestBlockCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /api/cards:
    get:
      tags:
        - card-rest-controller
      summary: Найти карты авторизированного пользователя
      description: Найти карты пользователя с пагинацией и фильтрацией
      operationId: getMyCards
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - BLOCK_REQUESTED
              - EXPIRED
        - name: minBalance
          in: query
          required: false
          schema:
            type: number
        - name: maxBalance
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardDTO"
  /api/cards/{cardId}/balance:
    get:
      tags:
        - card-rest-controller
      summary: Баланс карты
      description: Получить баланс карты пользователя по ID
      operationId: findCardBalance
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBalanceResponseDTO"
  /api/admin/users/{userId}/cards/{cardId}:
    delete:
      tags:
        - admin-rest-controller
      summary: Удалить пользователю карту
      operationId: deleteUserCard
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
components:
  schemas:
    UserUpdateDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        roleId:
          type: integer
          format: int64
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        roleId:
          type: integer
          format: int64
    CardUpdateDTO:
      type: object
      properties:
        expirationDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - BLOCK_REQUESTED
            - EXPIRED
        balance:
          type: number
    CardDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        ownerId:
          type: integer
          format: int64
        expirationDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - BLOCK_REQUESTED
            - EXPIRED
        balance:
          type: number
    AuthRequestDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    TransactionRequestDTO:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
    TransactionResultDTO:
      type: object
      properties:
        fromCardBalance:
          type: number
        toCardBalance:
          type: number
        status:
          type: string
          enum:
            - PENDING
            - SUCCESS
            - FAILED
    UserCreateDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        roleId:
          type: integer
          format: int64
    CardAddDTO:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
    CardCreateDTO:
      type: object
      properties:
        number:
          type: string
        expirationDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - BLOCK_REQUESTED
            - EXPIRED
        balance:
          type: number
    PageCardDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
    CardBalanceResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        balance:
          type: number
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT