# src/main/java/com/example/bankcards/dto/card/CardCreateDTO.java
package com.example.bankcards.dto.card;

import com.example.bankcards.entity.CardStatus;
import lombok.Getter;
import lombok.Setter;

import java.math.BigDecimal;
import java.time.LocalDate;

@Getter
@Setter
public class CardCreateDTO {

    private String number;
    private LocalDate expirationDate;
    private CardStatus status;
    private BigDecimal balance;
}