# src/main/java/com/example/bankcards/dto/card/CardDTO.java
package com.example.bankcards.dto.card;

import com.example.bankcards.entity.CardStatus;
import com.fasterxml.jackson.annotation.JsonFormat;
import lombok.Getter;
import lombok.Setter;

import java.math.BigDecimal;
import java.time.LocalDate;

@Getter
@Setter
public class CardDTO {

    private Long id;
    private String maskedNumber;
    private Long ownerId;
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
    private LocalDate expirationDate;
    private CardStatus status;
    private BigDecimal balance;

}